FILE: moves type definitions into separate folder
