+++ libs/spirit-islander/shared/util/src/index.ts
@@ -1,10 +1,10 @@
-export * from './lib/adversaries';
-export * from './lib/boards';
-export * from './lib/difficulty';
-export * from './lib/expansions';
-export * from './lib/maps';
-export * from './lib/option';
+export * from './lib/types/adversaries';
+export * from './lib/types/boards';
+export * from './lib/types/difficulty';
+export * from './lib/types/expansions';
+export * from './lib/types/maps';
+export * from './lib/types/option';
 export * from './lib/options';
-export * from './lib/players';
-export * from './lib/scenarios';
-export * from './lib/spirits';
+export * from './lib/types/players';
+export * from './lib/types/scenarios';
+export * from './lib/types/spirits';
