+++ libs/spirit-islander/shared/util/src/lib/data/maps.ts
@@ -1,4 +1,4 @@
-import { ExpansionName } from '../expansions';
+import { ExpansionName } from '../types/expansions';
 
 export const MAPS = [
   {
