+++ libs/spirit-islander/shared/util/src/lib/types/difficulty.ts
@@ -1,3 +1,3 @@
-import { DIFFICULTIES } from './data';
+import { DIFFICULTIES } from '../data';
 
 export type Difficulty = (typeof DIFFICULTIES)[number];
