+++ libs/spirit-islander/shared/util/src/lib/types/expansions.ts
@@ -1,3 +1,3 @@
-import { EXPANSIONS } from './data';
+import { EXPANSIONS } from '../data';
 
 export type ExpansionName = (typeof EXPANSIONS)[number];
