+++ libs/spirit-islander/shared/util/src/lib/types/maps.ts
@@ -1,4 +1,4 @@
-import { MAPS } from './data';
+import { MAPS } from '../data';
 import { DifficultyOption, ExpansionOption } from './option';
 
 export interface Map
