+++ libs/spirit-islander/shared/util/src/lib/types/option.ts
