+++ libs/spirit-islander/shared/util/src/lib/types/players.ts
@@ -1,3 +1,3 @@
-import { PLAYERS } from './data';
+import { PLAYERS } from '../data';
 
 export type Players = (typeof PLAYERS)[number];
