+++ libs/spirit-islander/shared/util/src/lib/types/scenarios.ts
@@ -1,4 +1,4 @@
-import { SCENARIOS } from './data';
+import { SCENARIOS } from '../data';
 import { DifficultyOption, ExpansionOption } from './option';
 
 export interface Scenario
