+++ libs/spirit-islander/shared/util/src/lib/types/spirits.ts
@@ -1,4 +1,4 @@
-import { SPIRITS } from './data';
+import { SPIRITS } from '../data';
 import { ExpansionOption } from './option';
 
 export type Spirit = ExpansionOption<SpiritName>;
